Refactored song search to work with main songs endpoint
The search endpoint is now handled by the main songs endpoint instead of
a separate search url. This was done to simplify things and be more in
line with REST principles.

Also updated duration inputs to have max of 59 instead of 60 as that
could be confusing. I.E. 5 mins and 60 secs is really 6 min